+++ test/production/app-dir/graceful-degrade-error/graceful-degrade-error-non-bot.test.ts
@@ -1,0 +1,39 @@
+// Duplicate the test file of graceful-degrade-error-bot.test.ts since we need to
+// restart with a new browser context for UA setting. Otherwise the browser context
+// will not be closed and reset properly. TODO: investigate why browser.close didn't help.
+
+import { nextTestSetup } from 'e2e-utils'
+import { deleteBrowserDynamicChunks } from './delete-dynamic-chunk'
+
+describe('graceful-degrade-error - non bot', () => {
+  const { next } = nextTestSetup({
+    files: __dirname,
+  })
+
+  // Delete client chunks to simulate chunk loading failure
+  beforeAll(() => {
+    deleteBrowserDynamicChunks(next)
+  })
+
+  it('should degrade to graceful error when chunk loading fails in ssr for bot', async () => {
+    const browser = await next.browser('/')
+
+    const logs = await browser.log()
+    const errors = logs
+      .filter((x) => x.source === 'error')
+      .map((x) => x.message)
+      .join('\n')
+
+    expect(errors).toMatch(/Failed to load resource./)
+    // Should not show the original content
+    const html = await browser.elementByCss('html')
+    const body = await browser.elementByCss('body')
+    expect(await html.getAttribute('class')).not.toBe('layout-cls')
+    expect(await body.getAttribute('class')).not.toBe('body-cls')
+
+    const bodyText = await body.text()
+    expect(bodyText).toMatch(
+      /Application error: a client-side exception has occurred while loading/
+    )
+  })
+})
