+++ test/production/app-dir/graceful-degrade-error/graceful-degrade-error.test.ts
@@ -1,6 +1,5 @@
-import fs from 'fs'
 import { nextTestSetup } from 'e2e-utils'
-import path from 'path'
+import { deleteBrowserDynamicChunks } from './delete-dynamic-chunk'
 
 describe('graceful-degrade-error', () => {
   const { next } = nextTestSetup({
@@ -9,27 +8,7 @@ describe('graceful-degrade-error', () => {
 
   // Delete client chunks to simulate chunk loading failure
   beforeAll(() => {
-    const clientChunkDir = path.join(next.testDir, '.next', 'static', 'chunks')
-    const clientChunkFiles = fs
-      .readdirSync(clientChunkDir)
-      // filter out the js file that contains the text "large test content"
-      .filter((filename) => {
-        const filePath = path.join(clientChunkDir, filename)
-        const isJsFile = filename.endsWith('.js')
-        const fileContent = isJsFile
-          ? fs.readFileSync(filePath, { encoding: 'utf8' })
-          : ''
-
-        return (
-          isJsFile && fileContent && fileContent.includes('large test content')
-        )
-      })
-      .map((file) => path.join(clientChunkDir, file))
-
-    // Intended to log to help debugging tests
-    console.log('Deleting client chunk files:', clientChunkFiles)
-    // delete all chunk files
-    clientChunkFiles.map((file) => fs.rmSync(file))
+    deleteBrowserDynamicChunks(next)
   })
 
   it('should degrade to graceful error when chunk loading fails in ssr for bot', async () => {
